fixing warnings

+++ cgpuvm-attest/src/lib.rs
@@ -12,10 +12,10 @@ pub fn attest(data: &[u8], pcrs: u32, endpoint_url: &str) -> Option<Vec<u8>> {
     unsafe {
         let url_ptr = endpoint_url_cstring.as_ptr();
         let mut dstlen = 32*1024;
-        let mut dst = Vec::with_capacity(dstlen as usize);
+        let mut dst = Vec::with_capacity(dstlen);
         let pdst = dst.as_mut_ptr();
         let res = get_attestation_token(data.as_ptr(), pcrs, pdst, &mut dstlen, url_ptr);
-        dst.set_len(dstlen as usize);
+        dst.set_len(dstlen);
         (res == 0).then_some(dst)
     }
 }
